+++ tests/test-facets.php
@@ -25,6 +25,9 @@ require_once __DIR__ . '/../includes/class-wmds-num.php';
 require_once __DIR__ . '/../includes/class-wmds-date.php';
 require_once __DIR__ . '/../includes/class-wmds-str.php';
 require_once __DIR__ . '/../includes/class-wmds-num.php';
+require_once __DIR__ . '/../includes/class-wmds-facet-request.php';
+require_once __DIR__ . '/../includes/class-wmds-facet-query.php';
+require_once __DIR__ . '/../includes/class-wmds-facet-store.php';
 require_once __DIR__ . '/../includes/class-wmds-facets.php';
 
 wmds_section( 'Reading a request' );